Temporally disabled AOE Avoiding.
Added a counter for number of attacks done and made it so after 12 attacks it will look for a new target. This should help prevent over chasing targets and many other things

Added an option to focus targeting Fallen Maniacs (the exploding assholes). I will be adding others as well over time.

Ensured that Sweeping Wind will only cast right before attacking a target instead of keeping it up all the time inside or outside combat.

So far it seems that we do not stall around anymore, at most only for a moment or so, then it gets right back at it just fine.

Made sure that all Focused targets are actually focused.

Minor code cleanup.