1/21/2013 Change Log
Added options to Dashing Strike so you can choose how close and how far the mobs have to be before you use Dashing Strike as well as how much spirit is neede before using it.

Added option to Seven-Sided Strike so you can choose to use it based on a HP % instead of just when there are enough mobs around to validate using it.

Changed the Focus target settings to false by default except for Treasure Seeker due to issues with getting stuck trying to run to them. I plan to have this fixed asap.

Increased the default scan distance for Elites to 60 from 40. That way we won't accidently miss them while goin from mob to mob.

Fixed Sweeping Wind not re-casting before it runs out. (added a OPTION to choose if you want to do this or not)

Moved MantraTimer to Mantra OPTIONS so people can choose to change how often it keeps the mantra up (remember CastMantraEvery3Seconds has to be set to true for it to be used)

Renamed fight_times to NumofAttacksTillChangeTarget (to be clearer on what it is for)

Moved meleeRange and NumofAttacksTillChangeTarget to the OPTIONS area and made sure they had comments next to them that was as clear as possible

Made some other tweaks on how close targets have to be before using some attack spells

Changed combatThreashold timer to 3 seconds from 10 seconds to help the rare times that we sit in place

Made some tweaks to how AOE Avoiding works. It is not a garantee to fix it at all, just steps in that direction.

Signed-off-by: Aaron Weiler <aaron@aaronweiler.net>