1/21 Changes
Worked on Mantra's Code a lot (read the following notes below marked with - at the beginning)
- Simplified the code by removing all of the Sweeping Wind checks and just having it be the overlapping if statement over everything to do with Mantras
- Removed all else's and just made each Mantra their own if statement to be easier to understand
- Added comments for each mantra so you can get an idea of what the big if statements do
- Re-named CastMantraEvery3Seconds to UseMantrasSecondaryBuff to be more clear on what it is for

Turned off AOE Avoiding by default to get more GPH and XP/H (just set AvoidAoE to true to turn it back on).
- Also did some minor work on AOE Avoiding, but it has been hell to get it to work as we want it to

Removed the meleeRange OPTION as it has became obsolite with all the new spell OPTIONS added over the last few months

Changed NumofAttacksTillChangeTarget to 3 from 10. (After every 3 attack spells used, it will look for the closest target again and set the current target to that closest target)

Moved the AOE and Spirit Regenerating attack spells to their own function

Made a function that uses any of the Spirit Regenerating Attack Spells that you have in front of the monk to get unstuck from that pack of mobs that are in his way to attack the target he wants to attack or to get to where he want to get to.

Added a check on the target's HP before and after the of attack spells. This way if we notice we have done no damage to the current target, we will first try to attack in front of our self, then regardless if that helped, we will change targets to whatever target is closest. This should really help make sure we do not sit and do nothing because the bot is trying to do something it cannot due to being stuck or whatever have you.

Started to make OPTIONS for all AOE spells so you can choose how far away the spell will be used on targets. So now you choose how far away the spell is used from). I also am making some OPTIONS on how many targets have to be within that set distance before using the spell. This way you get much more fine grain control over even more spells now. Allowing the use for almost any build imaginable with no issues in sight!